moved deposit/withraw tests into a perate file
